gateway-repo/app/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
email-validator==2.1.0
pika==1.3.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
redis==5.0.1
pytest==7.4.3
pytest-mock==3.12.0
prometheus_client==0.19.0